Resolved More Bugs (2)

+++ encrypt.go
@@ -87,7 +87,7 @@ func generateLTMatrix(matrixSize int) mat.Matrix {
 		// Generate Random slice
 		data := make([]float64, matrixSize*matrixSize)
 		for i := range data {
-			data[i] = float64(rand.Int())
+			data[i] = rand.Float64()
 		}
 
 		// Generate Random Matrix
